#258 Change styling for react-window list
This fixes janky loading of new elements when scrolling using a scrollwheel.
From https://github.com/bvaughn/react-window/issues/221
This only happens on Chrome, and the developer for react-window won't try to fix it but a solution suggested was removing the styling rule will-change if needed, so that's what I did.