Hotfix for multiple files and hidden window offloader
I wrote the code for PR #331 before PR #332 was decided on, and I tried to plan so my code would work with multiple LogViewer-components. I missed that props would be stale in the event listener, which is what was causing the problem. This hotfix fixes that problem, and the program now works again!