Hotfix for duplicate lines when a filter is set
The problem was that I was relying on the state in a hook to determine what to send, but update of the state is asynchronous.
I fixed this by using a ref that I update everytime the hook is called to be the latest length instead of using the state.